moved comment generator functions into separate file and used a setInterval to retrieve the comments continuously
