+++ blockly/tests/commentGenerator.js
@@ -1,3 +1,5 @@
+'use strict';
+
 /**
  * @license
  * Visual Blocks Editor
@@ -20,51 +22,187 @@
 
 /**
  * @fileoverview object that retrieves an array of blocks in the workspace and generates comments.
- * @author Alex Bowen
+ * @author Alex Bowen, Amber Libby
  */
 
-goog.provide('Blockly.Workspace');
-goog.require('goog.math');
 
-/**
-* function that retrieves the array of blocks within the workspace
-*/
-Blockly.CommentGenerator.prototype.getArray = function() {
-	var allBlocks = Blockly.Workspace.prototype.getAllBlocks;
-	return allBlocks;
-};
+function getImportantBlocks(){
+	//check if the workspace is empty
+	if (!xmlDoc || !xmlDoc.getElementsByTagName('BLOCK')) {
+		console.log("nothings here");
+        return null;
+    }
+    //add all blocks to the blockArr
+     var blockArr = xmlDoc.getElementsByTagName('BLOCK');
 
-/**
-* function that generates the comment structure
-*/
-Blockly.CommentGenerator.prototype.generateComments = function() {
-	var blockArray = getArray;
-	var allComments = null;
-	for (var i = 0; i <= blockArray; i++) {
-		var currentComment = blockArray[i].getCommentFromXML();
-		if(blockArray[i].hasComment == True) {
-			//ask luna how to identify if the block is nested or not.
-			if(blockArray[i].isTopBlock(id) == True) {
-				//pass block to comment area with <ul> or <span>
-				//ex.) allComments.add("<ul>"This is the comment in the block"</ul>")
-			};
-			else {
-				//check for how deeply nested the index is and then give it an <li> or <p>
-				//to match up with their <ul> or <span>
-			};
-		};
-	};
-	return allComments;
-};
+    var perfectArr = [];
 
-/**
-* function that will return true if the block at the given id has a comment or false if it doesn't
-*/
-Blockly.CommentGenerator.prototype.hasComment = function(id) {
-	if(block.getID(id).comment != null) {
-		return true;
-	};
-	else {
-		return false;
-	};
-};
+    //adding any blocks which can stand on their own to perfectArr
+    for(var i=0; i < blockArr.length; i++){
+
+		var strType = blockArr[i].getAttribute('type');
+
+		if(strType.match(/controls/g)){
+			perfectArr.push(blockArr[i]);
+		}
+		else if(strType.match(/procedures/g)){
+			perfectArr.push(blockArr[i]);
+		}
+		else if(strType == "beep"){
+			perfectArr.push(blockArr[i]);
+		}
+		else if(strType == "math_change") {
+			perfectArr.push(blockArr[i]);
+		}
+		else if(strType == "text_append") {
+			perfectArr.push(blockArr[i]);
+		}
+		else if(strType == "text_print") {
+			perfectArr.push(blockArr[i]);
+		}
+		else if(strType == "list_setIndex") {
+			perfectArr.push(blockArr[i]);
+		}
+		else if(strType == "variables_set") {
+			perfectArr.push(blockArr[i]);
+		}
+		else{
+			
+		}
+
+    }//end of for
+
+    getIndent(perfectArr);
+
+}//end of getImportantBlocks
+
+function getIndent(perfectArr){
+
+	//the string format of the current XML Doc
+	var currentXml = Blockly.Xml.domToPrettyText(Blockly.Xml.workspaceToDom(Blockly.mainWorkspace));
+	
+	var openStatementCnt;
+	var closeStatementCnt;
+	var indexOfId;
+	var idOfBlock;
+	var miniXml;
+	var i;
+	var parentArr = [];
+
+	for(i = 0; i < perfectArr.length; i++){
+
+		currentNode = perfectArr[i];
+		idOfBlock = currentNode.getAttribute('id');
+		indexOfId = currentXml.indexOf('id="'+idOfBlock+'"');
+		miniXml = currentXml.substring(0, indexOfId);
+		openStatementCnt = (miniXml.match(/<statement/g) || []).length;
+		closeStatementCnt = (miniXml.match(/statement>/g) || []).length;
+		parentArr[i] = openStatementCnt - closeStatementCnt;
+		parentArr.push(parentArr[i]);
+	}
+	parentArr.splice(i);
+	createComments(perfectArr, parentArr);
+
+}
+
+function commentPrefix(perfectArr, parentArr){
+	var zeroCount = 1;
+    var allCount = 0;
+    var prefixStringPrev;
+    var prefixArr = [];
+
+    for (var i = 0; i < parentArr.length; i++) {
+
+        if(parentArr[i].toString() == "0"){
+            prefixArr[i] = "C" + zeroCount.toString();
+            zeroCount++;
+            //console.log("this is level 0");
+        }
+        else{
+            var currentIndent = parentArr[i];
+            var prevIndent = parentArr[i-1];
+
+            if(currentIndent == prevIndent){
+                //at the same indent level
+
+                var shortStr = prefixArr[i-1].length-1;
+                var prevCount = prefixArr[i-1].substring(shortStr);
+
+                var newCount = parseInt(prevCount);
+                newCount++;
+
+                var preString = prefixArr[i-1].substring(0, shortStr);
+
+                prefixArr[i] = preString + newCount;
+            }
+            else if(currentIndent > prevIndent){
+                //there is another indent
+                prefixArr[i] = prefixArr[i-1] + ".1";
+                prefixStringPrev = prefixArr[i];
+            }
+            else if(currentIndent < prevIndent){
+                //there is one less indent here
+                var indentDiff = prevIndent - currentIndent;
+                var takeStr = indentDiff * 2;
+
+                var prevPrefixStr = prefixArr[i-1];
+                var prevPrefixStrLength = prevPrefixStr.length;
+                var subStrVal = prevPrefixStrLength - takeStr;
+
+                var thePrevPre = prevPrefixStr.substring(subStrVal);
+               //x is the prev prefix string name we need
+                var x = prevPrefixStr.substring(0, subStrVal-2);
+
+                //get the last char of the prev indent and up that value by 1 
+                //and append to the end of the prev string and set equal to the new prefix
+
+                var lastNum = prevPrefixStr.substring(subStrVal-1, subStrVal);
+                var num = parseInt(lastNum);
+                num++;
+                prefixArr[i] = x + "." + num;
+            }
+        }
+    }
+    return prefixArr;
+}
+
+function createComments(perfectArr, parentArr){
+  //clears the comment div of old data
+  document.getElementById("comment").innerHTML = "";
+
+  var pTag;
+  var commentStr;
+  var prefixes = commentPrefix(perfectArr, parentArr);
+  var indent;
+  for(var i = 0; i < perfectArr.length; i++){
+    commentStr = '';
+    currentNode = perfectArr[i];
+    pTag = document.createElement("p");
+    indent = parentArr[i];
+    //checks how many indents a comment is going to have
+    while(indent != 0) {
+      commentStr += "---";
+      indent--;
+    }
+    commentStr += " " + prefixes[i];
+
+    if(perfectArr[i].getElementsByTagName("comment")[0] == undefined){
+      commentStr += " No comment";
+    } 
+
+    else{
+        var parentsId = perfectArr[i].getElementsByTagName("comment")[0].parentNode.getAttribute('id');
+        if(parentsId == currentNode.getAttribute('id')){
+          var htmlComment = currentNode.getElementsByTagName("comment")[0].innerHTML;
+          commentStr += " " + htmlComment;
+        }
+        else{
+          commentStr += " No comment";
+        }
+    }
+
+    var pTextNode = document.createTextNode(commentStr);
+    pTag.appendChild(pTextNode);
+    document.getElementById("comment").appendChild(pTag);
+  }
+}
