+++ blockly/tests/navigation.js
@@ -16,30 +16,6 @@
 *limitations under the License.
 */
 
-goog.provide('Accessibility.Navigation');
-
-goog.require('Blockly.BlockSvg');
-goog.require('Blockly.Block');
-goog.require('Blockly.Connection');
-goog.require('Blockly.FieldColour');
-goog.require('Blockly.FieldTextInput');
-goog.require('Blockly.FieldDropdown');
-goog.require('Blockly.FieldCheckbox');
-
-goog.require('Blockly.Toolbox');
-
-goog.require('Blockly.Flyout');
-goog.require('goog.dom');
-goog.require('goog.events');
-goog.require('goog.events.BrowserFeature');
-goog.require('goog.html.SafeHtml');
-goog.require('goog.math.Rect');
-goog.require('goog.style');
-goog.require('goog.ui.tree.TreeControl');
-goog.require('goog.ui.tree.TreeNode');
-
-goog.require('Blockly.WidgetDiv');
-
 var xmlDoc = null;
 var currentNode = null;
 
@@ -567,147 +543,6 @@ function getBlockNodeById(id) {
     }
     // If you don't hit it return null.
     return null;
-}
-
-
-function getImportantBlocks(){
-	//check if the workspace is empty
-	if (!xmlDoc || !xmlDoc.getElementsByTagName('BLOCK')) {
-		console.log("nothings here");
-        return null;
-    }
-    //add all blocks to the blockArr
-     var blockArr = xmlDoc.getElementsByTagName('BLOCK');
-
-    var perfectArr = [];
-
-    //adding any blocks which can stand on their own to perfectArr
-    for(var i=0; i < blockArr.length; i++){
-
-		var strType = blockArr[i].getAttribute('type');
-
-		if(strType.match(/controls/g)){
-			perfectArr.push(blockArr[i]);
-		}
-		else if(strType.match(/procedures/g)){
-			perfectArr.push(blockArr[i]);
-		}
-		else if(strType == "beep"){
-			perfectArr.push(blockArr[i]);
-		}
-		else if(strType == "math_change") {
-			perfectArr.push(blockArr[i]);
-		}
-		else if(strType == "text_append") {
-			perfectArr.push(blockArr[i]);
-		}
-		else if(strType == "text_print") {
-			perfectArr.push(blockArr[i]);
-		}
-		else if(strType == "list_setIndex") {
-			perfectArr.push(blockArr[i]);
-		}
-		else if(strType == "variables_set") {
-			perfectArr.push(blockArr[i]);
-		}
-		else{
-			
-		}
-
-    }//end of for
-
-    getIndent(perfectArr);
-
-}//end of getImportantBlocks
-
-function getIndent(perfectArr){
-
-	//the string format of the current XML Doc
-	var currentXml = toXml();
-
-	var openStatementCnt;
-	var closeStatementCnt;
-	var indexOfId;
-	var idOfBlock;
-	var miniXml;
-	var i;
-	var parentArr = [];
-
-	for(i = 0; i < perfectArr.length; i++){
-
-		currentNode = perfectArr[i];
-		idOfBlock = currentNode.getAttribute('id');
-		indexOfId = currentXml.indexOf('id="'+idOfBlock+'"');
-		miniXml = currentXml.substring(0, indexOfId);
-		openStatementCnt = (miniXml.match(/<statement/g) || []).length;
-		closeStatementCnt = (miniXml.match(/statement>/g) || []).length;
-		parentArr[i] = openStatementCnt - closeStatementCnt;
-		parentArr.push(parentArr[i]);
-	}
-	parentArr.splice(i);
-	createComments(perfectArr, parentArr);
-
-}
-
-function commentPrefix(perfectArr, parentArr){
-	var zeroCount = 1;
-    var allCount = 0;
-    var prefixStringPrev;
-    var prefixArr = [];
-
-    for (var i = 0; i < parentArr.length; i++) {
-
-        if(parentArr[i].toString() == "0"){
-            prefixArr[i] = "C" + zeroCount.toString();
-            zeroCount++;
-            //console.log("this is level 0");
-        }
-        else{
-            var currentIndent = parentArr[i];
-            var prevIndent = parentArr[i-1];
-
-            if(currentIndent == prevIndent){
-                //at the same indent level
-
-                var shortStr = prefixArr[i-1].length-1;
-                var prevCount = prefixArr[i-1].substring(shortStr);
-
-                var newCount = parseInt(prevCount);
-                newCount++;
-
-                var preString = prefixArr[i-1].substring(0, shortStr);
-
-                prefixArr[i] = preString + newCount;
-            }
-            else if(currentIndent > prevIndent){
-                //there is another indent
-                prefixArr[i] = prefixArr[i-1] + ".1";
-                prefixStringPrev = prefixArr[i];
-            }
-            else if(currentIndent < prevIndent){
-                //there is one less indent here
-                var indentDiff = prevIndent - currentIndent;
-                var takeStr = indentDiff * 2;
-
-                var prevPrefixStr = prefixArr[i-1];
-                var prevPrefixStrLength = prevPrefixStr.length;
-                var subStrVal = prevPrefixStrLength - takeStr;
-
-                var thePrevPre = prevPrefixStr.substring(subStrVal);
-               //x is the prev prefix string name we need
-                var x = prevPrefixStr.substring(0, subStrVal-2);
-
-                //get the last char of the prev indent and up that value by 1 
-                //and append to the end of the prev string and set equal to the new prefix
-
-                var lastNum = prevPrefixStr.substring(subStrVal-1, subStrVal);
-                var num = parseInt(lastNum);
-                num++;
-                prefixArr[i] = x + "." + num;
-            }
-        }
-    }
-    return prefixArr;
 }
 
 //#endregion
